#72 - posting dto 패키지 하위에 request, response 추가 및 분류
