+++ src/main/java/com/ncookie/imad/domain/posting/dto/request/AddPostingRequest.java
@@ -1,6 +1,5 @@
-package com.ncookie.imad.domain.posting.dto;
+package com.ncookie.imad.domain.posting.dto.request;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
@@ -8,11 +7,12 @@ import lombok.Data;
 
 @Data
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class ModifyPostingRequest {
-    private String title;               // 게시글 제목
-    private String content;             // 게시글 본문
-    private int category;
+public class AddPostingRequest {
+    private Long contentsId;            // 작품 ID
 
-    @JsonProperty("is_spoiler")
+    private String title;               // 게시글 제목
+    private String content;             // 게시글  본문
+    private int category;               // 게시글 카테고리
+
     private boolean isSpoiler;          // 스포일러 여부
 }
