+++ src/main/java/com/ncookie/imad/domain/posting/dto/response/PostingDetailsResponse.java
@@ -1,4 +1,4 @@
-package com.ncookie.imad.domain.posting.dto;
+package com.ncookie.imad.domain.posting.dto.response;
 
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
