+++ src/main/java/com/ncookie/imad/domain/posting/service/PostingService.java
@@ -4,7 +4,11 @@ import com.ncookie.imad.domain.contents.entity.Contents;
 import com.ncookie.imad.domain.contents.service.ContentsService;
 import com.ncookie.imad.domain.like.entity.PostingLike;
 import com.ncookie.imad.domain.like.service.PostingLikeService;
-import com.ncookie.imad.domain.posting.dto.*;
+import com.ncookie.imad.domain.posting.dto.request.AddPostingRequest;
+import com.ncookie.imad.domain.posting.dto.request.ModifyPostingRequest;
+import com.ncookie.imad.domain.posting.dto.response.PostingDetailsResponse;
+import com.ncookie.imad.domain.posting.dto.response.PostingIdResponse;
+import com.ncookie.imad.domain.posting.dto.response.PostingListResponse;
 import com.ncookie.imad.domain.posting.entity.Posting;
 import com.ncookie.imad.domain.posting.repository.PostingRepository;
 import com.ncookie.imad.domain.user.entity.UserAccount;
